modules/gcp_infrastructure/python_scripts/function_code/bigquery/requirements.txt
async-timeout==4.0.2
backoff==2.2.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
Deprecated==1.2.13
google-api-core==2.11.0
google-auth==2.16.1
google-cloud==0.34.0
google-cloud-bigquery==3.5.0
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
idna==3.4
opentelemetry-api==1.15.0
opentelemetry-exporter-otlp-proto-grpc==1.15.0
opentelemetry-proto==1.15.0
opentelemetry-sdk==1.15.0
opentelemetry-semantic-conventions==0.36b0
packaging==23.0
proto-plus==1.22.2
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
redis==4.4.2
requests==2.28.2
rsa==4.9
six==1.16.0
typing_extensions==4.5.0
urllib3==1.26.14
wrapt==1.14.1
